Fix net-worth snapshot bounds, CSV import robustness, and repair the test suite
Three defects the test suite was already reporting, plus the reasons it could
not be trusted to say so.

Net worth:
- `AccountSnapshot.synced_at <= end` compared a datetime against a date, so
  "<= end" meant "<= midnight" and dropped every snapshot taken later in the
  day on the final day of the range. Since end_date is normally today, the
  chart always hid today's balances: the last point read 114,306.98 where the
  sidebar said 157,795.64. Now bounded by end + 1 day, exclusive.
- The guard against double-counting a source from both SourceBalanceHistory
  and AccountSnapshot also blocked *additional* snapshots of the same source,
  so a source was pinned to its earliest value and could never report
  history_mode "historical". It now excludes only sources already carried by
  history, which is what it was for.

Import: csv_importer caught (ValueError, KeyError), but Decimal("abc") raises
decimal.InvalidOperation, an ArithmeticError. One unparseable amount aborted
the whole file instead of skipping the row, which is what the comment claimed.

get_monthly_subscription_total returned an int when there was nothing to sum,
because round(0, 2) is 0.

The suite reported 61 failures, none of them product bugs beyond the above:

- tests/conftest.py did `from src.plugins.loader import _registry` and then
  called load_plugins(), which rebinds that module attribute. Every synthetic
  test plugin was appended to an orphaned list, so 16 tests exercised sources
  that did not exist.
- The CLI's serve/demo commands set FINANCE_APP_* in os.environ and
  Settings.load() treats those as overrides, so running the CLI tests before
  the config tests failed the config tests. An autouse fixture now restores
  the environment; the suite is order-independent.
- Patch targets that moved when imports went inside functions
  (src.main.init_db, .uvicorn, .Path), renamed config fields (plaid.env ->
  environment), and a required `currency` query param added later, so those
  422s were correct behaviour.
- Fixtures pinned to Feb 2026 that had aged out of the recurring detector's
  active window. Re-anchored to date.today() so they do not rot again.
- An earlier sanitization pass renamed test inputs but not the expectations
  they were matched against: a UBER.*EATS rule tested with EXAMPLE DELIVERY
  ORDER, a search for "Whole" against "Example Grocer", and merchant_raw
  "WAL-MART..." paired with merchant_clean "Example Store" in three dedupe
  tests, which defeated the merchant-similarity check every time.

Removed test_retirement_preview_marks_legacy_csv_duplicates: it patched
_legacy_retirement_source_ids, which no longer exists now that provider
scanning lives in plugins. Rewrote the pending-rows sync test, which asserted
a skip-pending branch that no longer exists — pending rows are stored and
promoted in place — so `skipped_pending` is now unreachable.

Also replaced two real brand names in public fixtures with synthetic ones.

Suite: 381 passed, 0 failed.

+++ src/processing/subscriptions.py
@@ -639,4 +639,6 @@ def detect_subscriptions(db: Session, rate_map: dict[str, float] | None = None) 
 
 
 def get_monthly_subscription_total(db: Session, rate_map: dict[str, float] | None = None) -> float:
-    return round(sum(item["monthly_equivalent"] for item in detect_subscriptions(db, rate_map=rate_map)), 2)
+    # float() matters: with nothing to sum, round(0, 2) is an int, so the return type
+    # changed shape depending on the data.
+    return float(round(sum(item["monthly_equivalent"] for item in detect_subscriptions(db, rate_map=rate_map)), 2))
